requirements.txt
pyaudio
adafruit-circuitpython-neopixel
numpy
recorder
scipy